Extracted getTimeUnit.
